langchain/news_research_tool_project/requirements.txt
# Core app
streamlit==1.38.0
python-dotenv==1.0.1

# LangChain (modern split)
langchain==0.3.7
langchain-openai==0.2.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2

# OpenAI SDK (modern 1.x)
openai==1.56.1
tiktoken==0.7.0

# Vector store
faiss-cpu==1.12.0  # Win + Python 3.12 compatible

# HTML/URL loading helpers for WebBaseLoader / Unstructured
unstructured==0.15.13
lxml==5.3.0
html2text==2024.2.26

# Windows-friendly libmagic shim
python-magic==0.4.27
python-magic-bin==0.4.14
